requirements.txt
streamlit>=1.33,<2
Pillow>=9
